Added Trigger script
Added Trigger script, which will perform certain actions to triggerable objects when triggered by a box collider or an InteractionController

The Trigger component requires a target, which is any script that inherits from Triggerable. When a script is chosen, certain parameters will be displayed. For instance, if a Door script is set as a target, you can tell that door to either open or close when the Trigger is triggered.

Some other triggerable scripts, such as a music player or a trigger when all enemies are killed, are to be added