Update MusicPlayer.cs
Added startPlay, soundtrack and segments fields.

Implemented the Play() and Stop() methods, along with the PlaySong() coroutine to play the soundtrack and the loop segments

The idea is to loop through song segments (because the music provider split the song loop into multiple files and I'm too lazy to combine them in Audacity) with the AudioSource soundtrack as an entry point..

Every AudioSource's loop and playOnAwake are disabled, as the script will take care of both situations.

If no loop segments are assigned, the AudioSource assigned on the soundtrack field will play once. If you want one entire AudioClip as a looping music, just put it as one loop segment and leave the soundtrack field blank.