Update MusicDirector.cs
Fully implemented the MusicDirector.

MusicDirector inherits Triggerable script, which will allow the use of Play() and Stop() methods from a Trigger

Added Dictionary that contains the name of the MusicPlayer's GameObject as the key and the MusicPlayer itself as the value.

In order to play a song, you need to pass the name of the song as parameter. Known song names are stored in the SongNames enum, which can be useful when you want to play a song from a Trigger. The MusicPlayer's GameObject's name must match its enum counterpart if you want to play it from a Trigger.