Update Level1.unity
Increased the global brightness as the map looked a little too dark.

Some of the lowpoly kitchen models had a weird emissive white material applied to them at randoms during light bake, so I made them all static in case that was the problem. I'll find out if that solved anything on the next bake.