Light bake
I know it's too late for this one, but light bakes wasted the last three hours jamming at the end and overall giving me a bad time. For whatever reason, the lamp posts aren't so bright anymore and the whole thing was almost ruined. At last this light bake is okay